src/app/modules/auth/services/guest.guard.ts
import { Injectable } from '@angular/core';
import { CanActivate, ActivatedRouteSnapshot, RouterStateSnapshot, Router } from '@angular/router';
import { Observable } from 'rxjs/Observable';
import { AuthService } from './auth.service';

@Injectable()
export class GuestGuard implements CanActivate {

    constructor(
        public authService: AuthService,
        public router: Router
    ) {}

    canActivate(
        next: ActivatedRouteSnapshot,
        state: RouterStateSnapshot
    ): boolean {
        if (this.authService.isLoggedIn) {
            this.router.navigate(['reservation/overview']);
            return false;
        }

        return true;
    }

}
